+++ ios/engine/KMEI/KeymanEngine/Classes/Settings/InstalledLanguagesViewController.swift
@@ -57,8 +57,8 @@ public class InstalledLanguagesViewController: UITableViewController, UIAlertVie
   
   override public func viewDidLoad() {
     super.viewDidLoad()
-    
-    title = "Installed Languages"
+
+    title = NSLocalizedString("menu-installed-languages-title", bundle: engineBundle, comment: "")
     selectedSection = NSNotFound
     packageDownloadStartedObserver = NotificationCenter.default.addObserver(
       forName: Notifications.packageDownloadStarted,
@@ -119,13 +119,13 @@ public class InstalledLanguagesViewController: UITableViewController, UIAlertVie
     
     // We do have updates - prepare the UI!
     if let toolbar = navigationController?.toolbar as? ResourceDownloadStatusToolbar {
-      toolbar.displayButton("Update available", with: self, callback: #selector(self.updateClicked))
+      toolbar.displayButton(NSLocalizedString("notification-update-available", bundle: engineBundle, comment: ""), with: self, callback: #selector(self.updateClicked))
     }
   }
   
   @objc func updateClicked(_ sender: Any) {
     if let toolbar = navigationController?.toolbar as? ResourceDownloadStatusToolbar {
-      toolbar.displayStatus("Updating\u{2026}", withIndicator: true)
+      toolbar.displayStatus(NSLocalizedString("notification-update-processing", bundle: engineBundle, comment: ""), withIndicator: true)
     }
     
     // Do the actual updates!
@@ -209,7 +209,11 @@ public class InstalledLanguagesViewController: UITableViewController, UIAlertVie
       cell.selectedBackgroundView = selectionColor
       cell.accessoryType = .disclosureIndicator
     }
-    cell.detailTextLabel?.text = keyboards.count < 2 ? (keyboards.first?.name ?? "") : "\(keyboards.count) installed"
+
+    let formatString = NSLocalizedString("settings-keyboards-installed-count", bundle: engineBundle, comment: "")
+    cell.detailTextLabel?.text = keyboards.count < 2 ?
+      (keyboards.first?.name ?? "") :
+      String.localizedStringWithFormat(formatString, keyboards.count)
     return cell
   }
   
@@ -275,31 +279,46 @@ public class InstalledLanguagesViewController: UITableViewController, UIAlertVie
       navigationItem.rightBarButtonItem?.isEnabled = true
     }
   }
-  
-  private func showDownloading(_ downloadLabel: String) {
+
+  private func packageDownloadStarted(key: KeymanPackage.Key) {
+    log.info("download started for \(key.type.rawValue): InstalledLanguagesViewController")
+    view.isUserInteractionEnabled = false
+    navigationItem.setHidesBackButton(true, animated: true)
+
     guard let toolbar = navigationController?.toolbar as? ResourceDownloadStatusToolbar else {
       return
     }
-    
-    toolbar.displayStatus("Downloading \(downloadLabel)\u{2026}", withIndicator: true)
-  }
 
-  private func packageDownloadStarted(key: KeymanPackage.Key) {
+    var msg: String
     switch(key.type) {
       case .keyboard:
-        keyboardDownloadStarted()
+        msg = NSLocalizedString("notification-downloading-keyboard", bundle: engineBundle, comment: "")
       case .lexicalModel:
-        lexicalModelDownloadStarted()
+        msg = NSLocalizedString("notification-downloading-lexical-model", bundle: engineBundle, comment: "")
     }
+
+    toolbar.displayStatus(msg, withIndicator: true)
   }
 
   private func packageDownloadCompleted(package: KeymanPackage) {
+    log.info("lexicalModelDownloadCompleted: InstalledLanguagesViewController")
+
+    var msg: String
+
     switch(package.resourceType()) {
       case .keyboard:
-        keyboardDownloadCompleted()
+        Manager.shared.shouldReloadKeyboard = true
+        msg = NSLocalizedString("notification-download-success-keyboard", bundle: engineBundle, comment: "")
       case .lexicalModel:
-        lexicalModelDownloadCompleted()
+        msg = NSLocalizedString("notification-download-success-lexical-model", bundle: engineBundle, comment: "")
     }
+
+    if let toolbar = navigationController?.toolbar as? ResourceDownloadStatusToolbar {
+      toolbar.displayStatus(msg, withIndicator: false, duration: 3.0)
+    }
+
+    restoreNavigation()
+    navigationController?.popToRootViewController(animated: true)
   }
 
   private func packageDownloadFailed(notification: PackageDownloadFailedNotification) {
@@ -307,76 +326,23 @@ public class InstalledLanguagesViewController: UITableViewController, UIAlertVie
       return
     }
 
-    switch(packageKey.type) {
-      case .keyboard:
-        keyboardDownloadFailed()
-      case .lexicalModel:
-        lexicalModelDownloadFailed()
-    }
-  }
-  
-  private func keyboardDownloadStarted() {
-    log.info("keyboardDownloadStarted: InstalledLanguagesViewController")
-    view.isUserInteractionEnabled = false
-    navigationItem.setHidesBackButton(true, animated: true)
-    showDownloading("keyboard")
-  }
-  
-  private func lexicalModelDownloadStarted() {
-    log.info("lexicalModelDownloadStarted")
-    view.isUserInteractionEnabled = false
-    navigationItem.setHidesBackButton(true, animated: true)
-    showDownloading("dictionary")
-  }
-  
-  private func keyboardDownloadCompleted() {
-    log.info("keyboardDownloadCompleted: InstalledLanguagesViewController")
-    Manager.shared.shouldReloadKeyboard = true
-    
-    if let toolbar = navigationController?.toolbar as? ResourceDownloadStatusToolbar {
-      toolbar.displayStatus("Keyboard successfully downloaded!", withIndicator: false, duration: 3.0)
-    }
-    restoreNavigation()
-    
-    navigationController?.popToRootViewController(animated: true)
-  }
-  
-  private func lexicalModelDownloadCompleted() {
-    log.info("lexicalModelDownloadCompleted: InstalledLanguagesViewController")
-    
-    if let toolbar = navigationController?.toolbar as? ResourceDownloadStatusToolbar {
-      toolbar.displayStatus("Dictionary successfully downloaded!", withIndicator: false, duration: 3.0)
-    }
-    
-    restoreNavigation()
-    navigationController?.popToRootViewController(animated: true)
-  }
-  
-  private func keyboardDownloadFailed() {
     log.info("keyboardDownloadFailed: InstalledLanguagesViewController")
 
-    if let toolbar = navigationController?.toolbar as? ResourceDownloadStatusToolbar {
-      toolbar.displayStatus("Download unsuccessful", withIndicator: false, duration: 3.0)
+    var msg: String
+    switch(packageKey.type) {
+      case .keyboard:
+        msg = NSLocalizedString("notification-download-failure-keyboard", bundle: engineBundle, comment: "")
+      case .lexicalModel:
+        msg = NSLocalizedString("notification-download-failure-lexical-model", bundle: engineBundle, comment: "")
     }
+
+    if let toolbar = navigationController?.toolbar as? ResourceDownloadStatusToolbar {
+      toolbar.displayStatus(msg, withIndicator: false, duration: 3.0)
+    }
+
     restoreNavigation()
 
     navigationController?.popToRootViewController(animated: true)
-  }
-  
-  private func lexicalModelDownloadFailed() {
-    log.info("lexicalModelDownloadFailed: InstalledLanguagesViewController")
-    restoreNavigation()
-    
-    let title = "Dictionary Download Error"
-    navigationController?.setToolbarHidden(true, animated: true)
-    
-    let alertController = UIAlertController(title: title, message: "",
-                                            preferredStyle: UIAlertController.Style.alert)
-    alertController.addAction(UIAlertAction(title: "OK",
-                                            style: UIAlertAction.Style.cancel,
-                                            handler: nil))
-    
-    self.present(alertController, animated: true, completion: nil)
   }
   
   private func batchUpdateStarted(_: [AnyLanguageResource]) {
@@ -388,15 +354,19 @@ public class InstalledLanguagesViewController: UITableViewController, UIAlertVie
       return
     }
     
-    toolbar.displayStatus("Updating\u{2026}", withIndicator: true)
+    toolbar.displayStatus(NSLocalizedString("notification-update-available", bundle: engineBundle, comment: ""), withIndicator: true)
   }
 
   private func batchUpdateCompleted(results: BatchUpdateCompletedNotification) {
     if let toolbar = navigationController?.toolbar as? ResourceDownloadStatusToolbar {
       if results.failures.count == 0 {
-        toolbar.displayStatus("\(results.successes.count) update(s) successfully downloaded!", withIndicator: false, duration: 3.0)
+        let formatString = NSLocalizedString("notification-update-success", bundle: engineBundle, comment: "")
+        toolbar.displayStatus(String.localizedStringWithFormat(formatString, results.successes.count), withIndicator: false, duration: 3.0)
       } else {
-        toolbar.displayStatus("Updates complete: \(results.successes.count) successful, \(results.failures.count) failed", withIndicator: false, duration: 3.0)
+        // Needs to be a short string on iPhones.
+        // Possible TODO:  condition on "is an iPad", provide more info on iPads?
+        let formatString = NSLocalizedString("notification-update-failed", bundle: engineBundle, comment: "")
+        toolbar.displayStatus(String.localizedStringWithFormat(formatString, results.failures.count), withIndicator: false, duration: 3.0)
       }
     }
     
@@ -454,14 +424,7 @@ public class InstalledLanguagesViewController: UITableViewController, UIAlertVie
   
   private func showConnectionErrorAlert() {
     dismissActivityView()
-    let alertController = UIAlertController(title: "Connection Error",
-                                            message: "Could not reach Keyman server. Please try again later.",
-                                            preferredStyle: UIAlertController.Style.alert)
-    alertController.addAction(UIAlertAction(title: "OK",
-                                            style: UIAlertAction.Style.default,
-                                            handler: errorAcknowledgmentHandler))
-    
-    self.present(alertController, animated: true, completion: nil)
+    Alerts.showDownloadErrorAlert(in: self, handler: errorAcknowledgmentHandler)
   }
 }